Add latest case card to index page
