+++ src/locales/zh/translation.json
@@ -27,6 +27,8 @@
   "waiting_time.last_updated": "最後更新：",
   "waiting_time.hour": "小時",
   "index.title": "疫情追蹤",
+  "index.latest_case": "最新個案",
+  "index.see_more": "查看更多",
   "index.highlight": "確診個案概要（{{count}}宗個案）",
   "dodgy_shops.title": "懷疑商鋪",
   "news.title": "即時新聞",
